@@ -1,6 +1,6 @@
 package com.samajackun.sumascript.parser;
 
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.fail;
 
 import java.io.IOException;
 import java.io.StringReader;
