@@ -1,11 +1,11 @@
 package com.samajackun.sumascript.parser;
 
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.fail;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import com.samajackun.rodas.core.model.NumericConstantExpression;
 import com.samajackun.sumascript.core.Instruction;
